requirements.txt
ccxt==4.5.42
pyyaml==6.0.2
pandas==2.2.3
sqlalchemy==2.0.36
apscheduler==3.10.4
streamlit==1.41.1
plotly==5.24.1
python-dotenv==1.0.1
